Proyect veterinarIA
I am seeking support to build a global strategy for non-profit veterinary assistance that operates through artificial intelligence, with the ability to provide support for health problems of pets, refer users to nearby clinics and offices, offer medications and products at a lower cost thanks to sponsorships received from large and medium-sized companies that can be linked to the strategy through advertising. With the support of special gadgets, it will be able to monitor vital signs and perform more precise online diagnoses.